refactor: Better architecture, QoL
