+++ client/package.json
@@ -22,9 +22,11 @@
     "@vitejs/plugin-react": "^3.1.0",
     "eslint-config-react-app": "^7.0.1",
     "vite": "^4.1.4",
-    "vite-plugin-node-polyfills": "^0.7.0"
+    "vite-plugin-node-polyfills": "^0.7.0",
+    "vite-tsconfig-paths": "^4.0.7"
   },
   "scripts": {
+    "start:dev": "vite build --watch",
     "start": "vite",
     "build": "tsc && vite build",
     "preview": "vite preview"
