+++ client/src/api/neighborhood-provider.tsx
@@ -1,19 +1,17 @@
 import { createContext, ReactNode, useState } from "react";
 
-const DEFAULT_NEIGHBORHOOD_ID = 1;
+const ID = 1;
 
 export const NeighborhoodContext = createContext({
-  id: DEFAULT_NEIGHBORHOOD_ID,
+  id: ID,
   setId: (_: number) => {},
 });
 
 export function NeighborhoodProvider({ children }: { children: ReactNode }) {
-  const [neighborhoodId, setNeighborhoodId] = useState(DEFAULT_NEIGHBORHOOD_ID);
+  const [id, setId] = useState(ID);
 
   return (
-    <NeighborhoodContext.Provider
-      value={{ id: neighborhoodId, setId: setNeighborhoodId }}
-    >
+    <NeighborhoodContext.Provider value={{ id, setId }}>
       {children}
     </NeighborhoodContext.Provider>
   );
