--- client/src/features/neighborhoods/components/index.ts
@@ -1,3 +1,0 @@
-export * from "./provider";
-export * from "./selector";
-export * from "./status";
