+++ client/src/ui/error-alert.tsx
