--- client/src/views/index.ts
@@ -1,8 +1,0 @@
-export { App } from "./app";
-export { AppProvider } from "./app.provider";
-export { Block } from "./block";
-export { Blocks } from "./blocks";
-export { Home } from "./home";
-export { Layout } from "./layout";
-export { Transaction } from "./transaction";
-export { Transactions } from "./transactions";
