+++ client/vite.config.ts
@@ -1,8 +1,14 @@
-import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react";
+import { defineConfig, splitVendorChunkPlugin } from "vite";
 import { nodePolyfills } from "vite-plugin-node-polyfills";
+import tsconfigPaths from "vite-tsconfig-paths";
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [react(), nodePolyfills({ protocolImports: true })],
+  plugins: [
+    react(),
+    nodePolyfills({ protocolImports: true }),
+    splitVendorChunkPlugin(),
+    tsconfigPaths(),
+  ],
 });
